streamlit_whisper/requirements.txt
chromedrivermanager==0.0.1
ipython==8.16.1
openpyxl==3.1.2
pip==23.2.1
pipdeptree==2.13.0
PyAutoGUI==0.9.54
PySocks==1.7.1
selenium==4.13.0
setuptools==68.2.0
streamlit==1.27.2
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
wheel==0.41.2
yt-whisper==1.0
python-docx==1.0.0




